orchestrator/requirements.txt
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
supabase>=2.3.0
docker>=7.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
tiktoken>=0.5.0
tenacity>=8.2.0
